fixed search ingredients by entry in field name + delete avatar implemention fixed

+++ backend/api/views.py
@@ -5,22 +5,32 @@ from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework.generics import (
     RetrieveUpdateDestroyAPIView,
     ListAPIView,
-    ListCreateAPIView,
     get_object_or_404,
-    ValidationError
+    ValidationError,
 )
 from rest_framework.views import APIView
 from rest_framework.mixins import CreateModelMixin, DestroyModelMixin
 from rest_framework.viewsets import GenericViewSet, ModelViewSet, ReadOnlyModelViewSet
 from django.contrib.auth import get_user_model
-from .serializers import UserSerializer, AvatarSerializer, SubscriptionSerializer, SubscribeSerializer, RecipeListOrRetrieveSerializer, RecipePostOrPatchSerializer, TagSerializer, GetOrRetriveIngredientSerializer, FavoriteSerializer, ShoppingCartSerializer
+from .serializers import (
+    UserSerializer,
+    AvatarSerializer,
+    SubscriptionSerializer,
+    SubscribeSerializer,
+    RecipeListOrRetrieveSerializer,
+    RecipePostOrPatchSerializer,
+    TagSerializer,
+    GetOrRetriveIngredientSerializer,
+    FavoriteSerializer,
+    ShoppingCartSerializer,
+)
 from core.models import Subscription, FavoriteRecipe
 from django.core.exceptions import PermissionDenied
 from django.http import Http404
 from django import urls
 from django.shortcuts import redirect
 from django_filters.rest_framework import DjangoFilterBackend
-from .filters import RecipeFilter, Recipe
+from .filters import IngredientFilter, Recipe, RecipeFilterSet
 from core.models import Recipe, ShortenedRecipeURL, Ingredient, Tag, ShoppingCart
 from .serializers import RecipeLinkSerializer
 
@@ -46,8 +56,17 @@ class UserAvatarUpdateView(RetrieveUpdateDestroyAPIView):
 
         if serializer.is_valid():
             serializer.save()
-            return Response({'status': 'Avatar updated'}, status=status.HTTP_200_OK)
+            return Response({"status": "Avatar updated"}, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, *args, **kwargs):
+        try:
+            user: User = get_object_or_404(User, pk=self.request.user.id)
+            user.avatar = None
+            user.save()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        except Exception as e:
+            return Response(e, status=status.HTTP_400_BAD_REQUEST)
 
 
 class SubscriptionsListView(ListAPIView):
@@ -63,7 +82,9 @@ class SubscriptionsListView(ListAPIView):
         subscriptions = Subscription.objects.filter(user=user)
 
         # Get the users who have subscribed to the current user
-        subscribed_users = User.objects.filter(id__in=subscriptions.values('subscribed_to'))
+        subscribed_users = User.objects.filter(
+            id__in=subscriptions.values("subscribed_to")
+        )
 
         # Return the queryset or an empty queryset if no subscriptions exist
         return subscribed_users if subscriptions.exists() else User.objects.none()
@@ -73,35 +94,43 @@ class SubscribeCreateDestroyView(CreateModelMixin, DestroyModelMixin, GenericVie
     serializer_class = SubscribeSerializer
 
     def get_object(self):
-        pk = self.kwargs.get('pk')
+        pk = self.kwargs.get("pk")
         subscribed_to = get_object_or_404(User, pk=pk)
-        return get_object_or_404(Subscription, user=self.request.user, subscribed_to=subscribed_to)
+        return get_object_or_404(
+            Subscription, user=self.request.user, subscribed_to=subscribed_to
+        )
 
     def perform_create(self, serializer):
-        pk = self.kwargs.get('pk')
+        pk = self.kwargs.get("pk")
         subscribed_to = get_object_or_404(User, pk=pk)
 
         # Check if the subscription already exists
         if self.request.user == subscribed_to:
             raise ValidationError("You cannot subscribe to yourself.")
 
-        if Subscription.objects.filter(user=self.request.user, subscribed_to=subscribed_to).exists():
+        if Subscription.objects.filter(
+            user=self.request.user, subscribed_to=subscribed_to
+        ).exists():
             raise ValidationError("You are already subscribed to this user.")
 
         serializer.save(user=self.request.user, subscribed_to=subscribed_to)
 
     def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data={})  # Empty data since `subscribed_to` is set internally
+        serializer = self.get_serializer(
+            data={}
+        )  # Empty data since `subscribed_to` is set internally
         serializer.is_valid(raise_exception=True)
         try:
             self.perform_create(serializer)
             headers = self.get_success_headers(serializer.data)
 
-        # Get the `subscribed_to` user instance
+            # Get the `subscribed_to` user instance
             subscribed_to_user = serializer.instance.subscribed_to
 
-        # Serialize the `subscribed_to` user using the `UserDetailSerializer`
-            user_detail_serializer = SubscriptionSerializer(subscribed_to_user, context={'request': request})
+            # Serialize the `subscribed_to` user using the `UserDetailSerializer`
+            user_detail_serializer = SubscriptionSerializer(
+                subscribed_to_user, context={"request": request}
+            )
             data = user_detail_serializer.data
 
             return Response(data, status=status.HTTP_201_CREATED, headers=headers)
@@ -115,87 +144,99 @@ class SubscribeCreateDestroyView(CreateModelMixin, DestroyModelMixin, GenericVie
 
 class RecipeViewSet(ModelViewSet):
     queryset = Recipe.objects.all()
-    # serializer_class = RecipeListOrRetrieveSerializer
-
-    # В зависимости от действия (list или retrieve) используем разные сериализаторы
-    def get_serializer_class(self):
-        if self.action in ['list', 'retrieve']:
-            return RecipeListOrRetrieveSerializer
-        elif self.action in ['create', 'update', 'partial_update']:
-            return RecipePostOrPatchSerializer
-
     filter_backends = [DjangoFilterBackend, filters.SearchFilter]
-    filterset_fields = [
-        'author', 'tags',
-        # 'is_favorited', 'is_in_shopping_cart'
-    ]
+    filterset_fields = ["author", "tags"]
+    filterset_class = RecipeFilterSet
+
+    def get_serializer_class(self):
+        if self.action in ["list", "retrieve"]:
+            return RecipeListOrRetrieveSerializer
+        elif self.action in ["create", "update", "partial_update"]:
+            return RecipePostOrPatchSerializer
+        return super().get_serializer_class()
 
 
 class RecipeLinkView(APIView):
     def get(self, request, id):
         try:
             recipe = Recipe.objects.get(pk=id)
-            if not hasattr(recipe, 'shortened_url'):
+            if not hasattr(recipe, "shortened_url"):
                 ShortenedRecipeURL.objects.create(recipe=recipe)
-            serializer = RecipeLinkSerializer(recipe, context={'request': request})
+            serializer = RecipeLinkSerializer(recipe, context={"request": request})
             return Response(serializer.data, status=status.HTTP_200_OK)
         except Recipe.DoesNotExist:
-            return Response({"detail": "Recipe not found."}, status=status.HTTP_404_NOT_FOUND)
+            return Response(
+                {"detail": "Recipe not found."}, status=status.HTTP_404_NOT_FOUND
+            )
 
 
 def redirect_to_original(request, short_code):
     url = get_object_or_404(ShortenedRecipeURL, short_code=short_code)
-    return redirect(urls.reverse('api:recipe-detail', kwargs={'pk': url.recipe.id}))
+    return redirect(urls.reverse("api:recipe-detail", kwargs={"pk": url.recipe.id}))
+
 
 class IngredientViewSet(ReadOnlyModelViewSet):
     queryset = Ingredient.objects.all()
     serializer_class = GetOrRetriveIngredientSerializer
+    # filter_backends = (DjangoFilterBackend,)
+    filterset_class = IngredientFilter
+    pagination_class = None
+
 
 class TagViewSet(ReadOnlyModelViewSet):
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
+    pagination_class = None
 
 
 class FavoriteView(ModelViewSet):
     queryset = FavoriteRecipe.objects.all()
     serializer_class = FavoriteSerializer
-    http_method_names = ['post', 'delete']
+    http_method_names = ["post", "delete"]
 
     def create(self, request, *args, **kwargs):
-        recipe_id = self.kwargs.get('id')  # Get recipe_id from URL
+        recipe_id = self.kwargs.get("id")  # Get recipe_id from URL
         recipe = get_object_or_404(Recipe, id=recipe_id)  # Retrieve the Recipe instance
 
-        data = {'user': request.user.id, 'recipe': recipe.id}
+        data = {"user": request.user.id, "recipe": recipe.id}
 
         serializer = self.get_serializer(data=data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
 
         headers = self.get_success_headers(serializer.data)
-        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+        return Response(
+            serializer.data, status=status.HTTP_201_CREATED, headers=headers
+        )
 
     def destroy(self, request, *args, **kwargs):
-        recipe_id = self.kwargs.get('id')  # Get recipe_id from URL
+        recipe_id = self.kwargs.get("id")  # Get recipe_id from URL
         recipe = get_object_or_404(Recipe, id=recipe_id)  # Retrieve the Recipe instance
 
         try:
-            favorite_recipe = FavoriteRecipe.objects.get(recipe=recipe, user=self.request.user)
+            favorite_recipe = FavoriteRecipe.objects.get(
+                recipe=recipe, user=self.request.user
+            )
             favorite_recipe.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         except FavoriteRecipe.DoesNotExist:
-            return Response({"detail": "This recipe is not in your favorites."}, status=status.HTTP_404_NOT_FOUND)
+            return Response(
+                {"detail": "This recipe is not in your favorites."},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+
 
 class ShoppingCartView(CreateModelMixin, DestroyModelMixin, GenericViewSet):
     queryset = ShoppingCart.objects.all()
     serializer_class = ShoppingCartSerializer
-    http_method_names = ['post', 'delete']
+    http_method_names = ["post", "delete"]
 
     def create(self, request, *args, **kwargs):
-        recipe_id = self.kwargs.get('id')
+        recipe_id = self.kwargs.get("id")
         recipe = get_object_or_404(Recipe, id=recipe_id)
 
         # Поскольку user передается в request и уже есть в контексте, добавим его к данным
-        data = {'user': request.user.id, 'recipe': recipe.id}
+        data = {"user": request.user.id, "recipe": recipe.id}
 
         serializer = self.get_serializer(data=data)
         serializer.is_valid(raise_exception=True)
@@ -204,15 +245,27 @@ class ShoppingCartView(CreateModelMixin, DestroyModelMixin, GenericViewSet):
         serializer.save(recipe=recipe, user=request.user)
 
         headers = self.get_success_headers(serializer.data)
-        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+        return Response(
+            serializer.data, status=status.HTTP_201_CREATED, headers=headers
+        )
 
     def destroy(self, request, *args, **kwargs):
-        recipe_id = self.kwargs.get('id')
+        recipe_id = self.kwargs.get("id")
         recipe = get_object_or_404(Recipe, id=recipe_id)
 
-        shopping_cart_item = ShoppingCart.objects.filter(recipe=recipe, user=request.user).first()
+        shopping_cart_item = ShoppingCart.objects.filter(
+            recipe=recipe, user=request.user
+        ).first()
+        if request.user.is_anonymous:
+            return Response(
+                {"detail": "You're unauthorized for this action."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         if shopping_cart_item:
             shopping_cart_item.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         else:
-            return Response({"detail": "This recipe is not in your shopping cart."}, status=status.HTTP_404_NOT_FOUND)
+            return Response(
+                {"detail": "This recipe is not in your shopping cart."},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
